horrible code, need a complete refactor

+++ machinehub/server/app/controllers/machine_controller.py
@@ -2,11 +2,11 @@ from flask.globals import request
 import os
 from flask_classy import route, FlaskView
 from flask.templating import render_template
-from machinehub.server.app.controllers.auth_controller import requires_auth
 from machinehub.config import UPLOAD_FOLDER, MACHINES_FOLDER, MACHINESOUT
 from machinehub.server.app.models.machine_model import MachineModel
 from machinehub.server.app.controllers.form_generator import metaform
 from machinehub.common.sha import dict_sha1
+from flask_login import login_required, current_user
 
 
 types = {'int': int,
@@ -24,6 +24,7 @@ class MachineController(FlaskView):
     @route('/<machine_name>', methods=['GET'])
     def machine(self, machine_name):
         show_stl = False
+        authoraize_user = self._user_is_owner(machine_name)
         if machine_name not in self.machines_model:
             return render_template('404.html'), 404
         doc, inputs = self.machines_model.machine(machine_name)
@@ -35,11 +36,13 @@ class MachineController(FlaskView):
                                form=form,
                                show_stl=show_stl,
                                file_name="",
-                               machine_name=machine_name)
+                               machine_name=machine_name,
+                               authoraize_user=authoraize_user)
 
     @route('/<machine_name>', methods=['POST'])
     def post_machine(self, machine_name):
         show_stl = False
+        authoraize_user = self._user_is_owner(machine_name)
         if machine_name not in self.machines_model:
             return render_template('404.html'), 404
         doc, inputs = self.machines_model.machine(machine_name)
@@ -71,11 +74,25 @@ class MachineController(FlaskView):
                                form=form,
                                show_stl=show_stl,
                                file_name=file_url,
-                               machine_name=machine_name)
+                               machine_name=machine_name,
+                               authoraize_user=authoraize_user)
 
     @route('/<machine_name>', methods=['DELETE'])
-    @requires_auth
+    @login_required
     def delete_machine(self, machine_name):
         if machine_name not in self.machines_model:
             return render_template('404.html'), 404
-        self.machines_model.delete(machine_name)
+        authoraize_user = self._user_is_owner(machine_name)
+        if authoraize_user:
+            from machinehub.server.app.models.user_model import UserMachine
+            from machinehub.server.app.webapp import db
+            machine = UserMachine(machine_name)
+            del machine
+            db.session.commit()
+            self.machines_model.delete(machine_name)
+
+    def _user_is_owner(self, machine_name):
+        authoraize_user = False
+        if current_user.is_authenticated():
+            authoraize_user = machine_name in [m.machinename for m in current_user.machines.all()]
+        return authoraize_user
